fix: abstract the checkStatus function, overdue added
