+++ packages/invoice-dashboard/src/lib/view-requests.svelte
@@ -31,6 +31,7 @@
   import type { IConfig } from "@requestnetwork/shared-types";
   import type { RequestNetwork } from "@requestnetwork/request-client.js";
   // Utils
+  import { checkStatus } from "@requestnetwork/shared-utils/checkStatus";
   import { config as defaultConfig } from "@requestnetwork/shared-utils/config";
   import { initializeCurrencyManager } from "@requestnetwork/shared-utils/initCurrencyManager";
   import { exportToPDF } from "@requestnetwork/shared-utils/generateInvoice";
@@ -38,7 +39,7 @@
   import { CurrencyManager } from "@requestnetwork/currency";
   import { onDestroy, onMount, tick } from "svelte";
   import { formatUnits } from "viem";
-  import { capitalize, debounce, formatAddress } from "../utils";
+  import { debounce, formatAddress } from "../utils";
   import { Drawer, InvoiceView } from "./dashboard";
   import { getPaymentNetworkExtension } from "@requestnetwork/payment-detection";
   import { CurrencyTypes } from "@requestnetwork/types";
@@ -380,32 +381,6 @@
 
   const handleRemoveSelectedRequest = () => {
     activeRequest = undefined;
-  };
-
-  const checkStatus = (request: Types.IRequestDataWithEvents | null) => {
-    if (Number(request?.balance?.balance)! > 0) {
-      return request?.balance?.balance! >= request?.expectedAmount!
-        ? "Paid"
-        : "Partially Paid";
-    }
-
-    const eventStatus = {
-      reject: "Rejected",
-      overdue: "Overdue",
-      cancel: "Canceled",
-    };
-
-    for (const [event, status] of Object.entries(eventStatus)) {
-      if (
-        request?.events?.some(
-          (e: { name?: string }) => e?.name?.toLowerCase() === event
-        )
-      ) {
-        return status;
-      }
-    }
-
-    return capitalize(request?.state as string);
   };
 </script>
