+++ shared/utils/capitalize.ts
@@ -1,0 +1,2 @@
+export const capitalize = (str: string) =>
+  (str && str[0].toUpperCase() + str.slice(1)) || "";
